backend/requirements.txt
fastapi
uvicorn[standard]
databricks-sdk
requests
psycopg2-binary
python-multipart
Pillow
